refactor: generated -> e2e
